From-SVN: r194912

+++ gcc/testsuite/gcc.target/mips/pr55315.c
@@ -5,7 +5,7 @@ int data[4096];
 int
 f (void)
 {
-  return (((unsigned int) &data[0]) == 0xdeadbea0U);
+  return (((unsigned long) &data[0]) == 0xdeadbea0U);
 }
 
 /* { dg-final { scan-assembler-not "\tmove\t\\\$2,\\\$0" } } */
